Move models from root.
